+++ kappadata/wrappers/dataset_wrappers/shuffle_wrapper.py
@@ -10,6 +10,6 @@ class ShuffleWrapper(KDSubset):
             self.rng = np.random.default_rng(seed=seed)
         else:
             self.rng = np.random
-        indices = np.arange(len(dataset))
+        indices = np.arange(len(dataset), dtype=np.int32)
         self.rng.shuffle(indices)
         super().__init__(dataset=dataset, indices=indices)
